dsrag/dsparse/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via instructor
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.37.1
    # via -r requirements.in
anyio==4.6.2.post1
    # via
    #   anthropic
    #   httpx
    #   openai
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via typer
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.16
    # via instructor
docx2txt==0.8
    # via -r requirements.in
exceptiongroup==1.2.2
    # via anyio
filelock==3.16.1
    # via huggingface-hub
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.10.0
    # via huggingface-hub
google-ai-generativelanguage==0.6.10
    # via google-generativeai
google-api-core[grpc]==2.21.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.149.0
    # via google-generativeai
google-auth==2.35.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.3
    # via -r requirements.in
googleapis-common-protos==1.65.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.67.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.67.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.27.2
    # via
    #   anthropic
    #   langsmith
    #   openai
huggingface-hub==0.26.1
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
instructor==1.6.3
    # via -r requirements.in
jinja2==3.1.4
    # via instructor
jiter==0.5.0
    # via
    #   anthropic
    #   instructor
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.12
    # via langchain-text-splitters
langchain-text-splitters==0.3.0
    # via -r requirements.in
langsmith==0.1.137
    # via langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
openai==1.52.2
    # via
    #   -r requirements.in
    #   instructor
orjson==3.10.10
    # via langsmith
packaging==24.1
    # via
    #   huggingface-hub
    #   langchain-core
pdf2image==1.17.0
    # via -r requirements.in
pillow==11.0.0
    # via
    #   -r requirements.in
    #   pdf2image
propcache==0.2.0
    # via yarl
proto-plus==1.25.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.28.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   anthropic
    #   google-generativeai
    #   instructor
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.23.4
    # via
    #   instructor
    #   pydantic
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via httplib2
pypdf2==3.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain-core
requests==2.32.3
    # via
    #   google-api-core
    #   huggingface-hub
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.3
    # via
    #   instructor
    #   typer
rsa==4.9
    # via google-auth
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
tenacity==9.0.0
    # via
    #   instructor
    #   langchain-core
tokenizers==0.20.1
    # via anthropic
tqdm==4.66.5
    # via
    #   google-generativeai
    #   huggingface-hub
    #   openai
typer==0.12.5
    # via instructor
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   google-generativeai
    #   huggingface-hub
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   pypdf2
    #   rich
    #   typer
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
yarl==1.16.0
    # via aiohttp
